PR Comments: Move more functions to unix_utils
